Cart: Remove `IsInitialized()`, add `GetType()`
`IsInitialized()` was only used by BackupMemoryCartridge because the constructor could "fail". Simplifying the constructor such that the burden of creation of the bup::BackupMemory lies on the hands of the user also simplifies `InsertCartridge()` methods.

`GetType()` enables fast type reflection -- it's now possible to get the concrete implementation with `GetType() -> static_cast<ConcreteType *>`

These should make it possible to retrieve the concrete type of the inserted cartridge at runtime, but thread-safety is still a concern