feat(data_ops): Remove `{Read,Write}{BE,LE}` explicit call to `memcpy`
These calls to `memcpy` are _actual_ calls to `memcpy` in debug-mode and can be replaced with simpler pointer-arithmetic to remove the overhead of these high-traffic data-ops.

What used to be a `memcpy`-call is now just a simple move-instruction.

`ldrh`/`strh` in the case of the arm64 device I am testing this on. Likely `mov` on x64 and maybe this even lends itself a bit better for the compiler to emit a [movbe](https://www.felixcloutier.com/x86/movbe) instruction.