feat(VDP): Use bool-arrays rather than bitsets
Seems to lend itself better to generating vectorized code with an execution mask, such as utilizing `vpmovzxbd` to quickly turn an array of bools into a vector-mask. Not so much on Arm64 but some specialized All/Any checks have been implemented to help with some of the bad codegen there.