fix(VDP): Replace `_mm_andnot_ps` with `_mm_andnot_si128`
These emit two different instructions. `andnps` vs `pandn`.
There is an additional latency penalty moving data between the integer-domain and the floating-point domain. We want the `pandn` form of this instruction and sustain our data in the integer-domain through-out.

See section 13 of agner's doc for more info:
https://www.agner.org/optimize/optimizing_assembly.pdf

```
There is a performance penalty for using the wrong type of instructions on some
processors. This is because the processor may have different data buses or different
execution units for integer and floating point data. Moving data between the integer and
floating point units can take one or more clock cycles depending on the processor, as listed
in table 13.3.```
```