fix(VDP): Fix `AllZeroU8` SSE2 bitmask check
`_mm_cmpeq_epi8` is comparing the input array to zero, which sets the `uint8`s to FF when the corresponding value is zero or 00 otherwise. `_mm_movemask_epi8` then condenses the array into an `uint16`. Since we want to check for all-zeros, the bitmask check should bail out if any bit is clear (non-zero) instead of set (zero).

Fixes black graphics everywhere on SSE2 builds