fix(savestates): Add SCILV to SCSP state and save/load `egAttackBug` flag to SCSP slot state
SCILV is a major problem as the data is completely missing from old versions of save states. A separate flag, `reuseSCILV`, was added to the state object in order to preserve the old behavior when loading save states from previous versions.

SCSP reuses the current SCILV values when loading states, which leads to broken audio under some circumstances when loading save states.