fix(app): Acquire Window-DPI rather than Display DPI
This allows the application to be more responsive to possible Multi-DPI setups or Per-Window DPI settings rather than using the Display Scaling of the primary display.

This fixes MacOS reporting a DPI of `1.0` despite the DPI being 2.0. This does not fix MacOS window-coordinates still being `1.0` though due to MacOS handling of Window coordinates between Logical Pixels and "Real" pixels.