feat(app): Automatically detect profile path and allow using the user profile path from the OS
Ymir tries to detect the profile path by locating Ymir.toml in the portable path
first and then in the user's home directory. If it can't locate Ymir.toml in one
of these directories, it will show a dialog on startup, in which the desired
profile mode can be selected. "Installed" mode uses the user's home directory
and "Portable" mode uses the portable path, i.e. the current working directory.

Specifying the -p option still overrides this behavior and uses the given custom
profile path.

A new option -u is available which forces usage of the user's home directory and
which overrides all other behavior.

Contains additional cosmetic code changes:
* put the profile selection code closer to where the settings file is loaded
* replace multiple occurrences of string literal "Ymir.toml" with a constant
* trim some trailing whitespaces